+++ src/QEPOptimize.jl
@@ -1,5 +1,7 @@
 module QEPOptimize
 
+import Base:+,==
+
 using BPGates
 using BPGates: mctrajectory!, continue_stat, PauliNoise, BellMeasure, CNOTPerm, toQCcircuit
 # TODO (mctrajectory!, continue_stat, PauliNoise) should be exported by default